Reorganize the pricing and connection folders
