+++ src/components/PricingPage/Faq/faqs/index.tsx
