+++ src/components/PricingPage/Faq/faqs/styles.module.less
